+++ src/main/java/com/example/demo/implement/tokenmanagerment/JwtTokenManager.java
@@ -1,4 +1,4 @@
-package com.example.demo.implement.auth;
+package com.example.demo.implement.tokenmanagerment;
 
 import static com.example.demo.exception.ExceptionType.*;
