+++ src/test/java/com/example/demo/implement/auth/JwtTokenManagerTest.java
@@ -11,6 +11,7 @@ import com.example.demo.dto.auth.JwtClaims;
 import com.example.demo.exception.ExceptionType;
 import com.example.demo.exception.JwtCustomException;
 import com.example.demo.factory.TestJwtFactory;
+import com.example.demo.implement.tokenmanagerment.JwtTokenManager;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
